+++ regimes/common/common.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -26,10 +27,13 @@ var (
 
 // NormalizeTaxIdentity removes any whitespace or separation characters and ensures all letters are
 // uppercase.
-func NormalizeTaxIdentity(tID *tax.Identity) error {
+func NormalizeTaxIdentity(tID *tax.Identity, altCodes ...l10n.Code) error {
 	code := strings.ToUpper(tID.Code.String())
 	code = TaxCodeBadCharsRegexp.ReplaceAllString(code, "")
 	code = strings.TrimPrefix(code, string(tID.Country))
+	for _, alt := range altCodes {
+		code = strings.TrimPrefix(code, string(alt))
+	}
 	tID.Code = cbc.Code(code)
 	return nil
 }