+++ regimes/gr/gr.go
@@ -6,6 +6,7 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
+	"github.com/invopop/gobl/l10n"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 )
@@ -31,7 +32,10 @@ const (
 // New provides the tax region definition
 func New() *tax.Regime {
 	return &tax.Regime{
-		Country:  "EL",
+		Country: "EL",
+		AltCountryCodes: []l10n.Code{
+			"GR", // regular ISO code
+		},
 		Currency: currency.EUR,
 		Name: i18n.String{
 			i18n.EN: "Greece",